One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/planning/MerchSetup/merchlist/merchlist.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { MerchlistComponent } from './merchlist.component';

describe('Component: Merchlist', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [MerchlistComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([MerchlistComponent],
      (component: MerchlistComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(MerchlistComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(MerchlistComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-merchlist></app-merchlist>
  `,
  directives: [MerchlistComponent]
})
class MerchlistComponentTestController {
}

